% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcmm.R
\name{calcmm}
\alias{calcmm}
\title{Calculate the minimum misclassification rate in a cross-tabulation}
\usage{
calcmm(tab)
}
\arguments{
\item{tab}{a cross-tabulation table of the simulation-based class solution
that resulted in maximum D to the true class solution}
}
\description{
This function provides a way of reconciling the arbitrary class
labels that result from k-means clustering to obtain the misclassification
rate of the k-means clustering result to the true class solution
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
