% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{d}
\alias{d}
\title{Estimate the incremental explained risk variation}
\usage{
d(formula, var, M, data)
}
\arguments{
\item{formula}{an \code{mlogit::mFormula()} model formula for a polytmous
logistic regression model to be fit with \code{mlogit()} using the
appropriate variable names from the data of interest}

\item{var}{the name of the subtype variable in the data, where 0 indicates
control subjects, should be supplied in quotes, e.g. var = "subtype"}

\item{M}{the number of subtypes. This should not include controls, but
only the number of subtypes among case subjects. For M>=2.}

\item{data}{the name of the dataframe that contains the relevant variables}
}
\description{
\code{d} estimates the incremental explained risk variation
across a set of pre-specified disease subtypes in a case-control study.
This function takes a model formula and a wide dataset and does the needed
transformation on the dataset to get the correct format, fits the polytomous
logistic regression model, and calculates D based on the resulting risk
predictions.
}
\examples{

# specify the model formula for three risk factors, x1, x2 and x3
mform <- mlogit::mFormula(subtype ~ 1 | x1 + x2 + x3)

# calculate D
d(mform, "subtype", 3, subtype_data)

}
\references{
Begg, C. B., Zabor, E. C., Bernstein, J. L., Bernstein, L., Press, M. F., &
Seshan, V. E. (2013). A conceptual and methodological framework for
investigating etiologic heterogeneity. Stat Med, 32(29), 5039-5052.
doi: 10.1002/sim.5902
}
