% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dest.R
\name{dest}
\alias{dest}
\title{Estimate the incremental explained risk variation}
\usage{
dest(formula, cls, M, data)
}
\arguments{
\item{formula}{an mFormula() model formula for a polytmous logistic
regression model to be fit with mlogit() using the appropriate variable
names from the data of interest, see Examples}

\item{cls}{the name of the subtype variable in the data, where 0 indicates
control subjects, should be supplied in quotes, e.g. cls = "class"}

\item{M}{the number of subtypes. This should not include controls, but
only the number of subtypes among case subjects. For M>=2.}

\item{data}{the name of the dataframe that contains the relevant variables}
}
\description{
\code{dest} estimates the incremental explained risk variation
across a set of pre-specified disease subtypes in a case-control study
}
\examples{

# generate data - 4 classes and 2 risk factors
class <- rep(c(0, 1, 2, 3, 4), times = c(1000, 250, 250, 250, 250))
x <- matrix(rnorm(2000 * 2), 2000, 2) +
model.matrix(~factor(class))[, -1] \%*\%
t(matrix(c(1.5, 0, 0.75, 0.25, 0.25, 0.75, 0, 1.5), ncol = 4))
df <- data.frame(class = class, x1 = x[, 1], x2 = x[, 2])

# specify the model formula
library(mlogit)
mform <- mFormula(class ~ 1 | x1 + x2)

dest(mform, "class", 4, df)

}
\references{
Begg, C. B., Zabor, E. C., Bernstein, J. L., Bernstein, L., Press, M. F., &
Seshan, V. E. (2013). A conceptual and methodological framework for
investigating etiologic heterogeneity. Stat Med, 32(29), 5039-5052.
doi: 10.1002/sim.5902
}
