% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstarest.R
\name{dstarest}
\alias{dstarest}
\title{Estimate the incremental explained risk variation in case-only study}
\usage{
dstarest(formula, cls, M, data)
}
\arguments{
\item{formula}{an mFormula() model formula for a polytmous logistic
regression model to be fit with mlogit() using the appropriate variable
names from the data of interest, see Examples}

\item{cls}{the name of the subtype variable in the data with values 1
through M. Should be supplied in quotes, e.g. cls = "class"}

\item{M}{the number of subtypes among case subjects. For M>=2.}

\item{data}{the name of the dataframe that contains the relevant variables}
}
\description{
\code{dstarest} estimates the incremental explained risk variation
across a set of pre-specified disease subtypes in a case-only study
}
\examples{

# generate data - 4 classes and 2 risk factors
set.seed(20180521)
class <- rep(c(0, 1, 2, 3, 4), times = c(1000, 250, 250, 250, 250))
x <- matrix(rnorm(2000 * 2), 2000, 2) +
model.matrix(~factor(class))[, -1] \%*\%
t(matrix(c(1.5, 0, 0.75, 0.25, 0.25, 0.75, 0, 1.5), ncol = 4))
df <- data.frame(class = class, x1 = x[, 1], x2 = x[, 2])

# remove the controls - this is a case only analysis example
df <- df[df$class != 0, ]

library(mlogit)
mform <- mFormula(class ~ 1 | x1 + x2)

dstarest(mform, "class", 4, df)

}
\references{
Begg, C. B., Seshan, V. E., Zabor, E. C., Furberg, H., Arora, A.,
Shen, R., . . . Hsieh, J. J. (2014). Genomic investigation of etiologic
heterogeneity: methodologic challenges. BMC Med Res Methodol, 14, 138.
doi: 10.1186/1471-2288-14-138
}
