% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehCalcD.R
\name{ehCalcD}
\alias{ehCalcD}
\title{Calculate the D metric to measure the extent of etiologic heterogeneity in
a case-control study}
\usage{
ehCalcD(data, cls, M, formula)
}
\arguments{
\item{data}{the name of the dataframe that contains the relevant variables}

\item{cls}{the names of the subtype variable in the data, should be
supplied in quotes, e.g. cls = "class"}

\item{M}{the number of subtypes. This could should not include controls, but
only the number of subtypes among case subjects.}

\item{formula}{an mFormula() model formula for a polytmous logistic
regression model to be fit with mlogit() using the appropriate variable
names from the data as interest e.g. formula = mFormula(class ~ 1 | x1 + x2)
for a model with subtype variable class and two individual-specific
predictors x1 and x2}
}
\description{
Calculate the D metric to measure the extent of etiologic heterogeneity in
a case-control study
}
