% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehpoly.R
\name{ehpoly}
\alias{ehpoly}
\title{Conduct an analysis of etiologic heterogeneity using polytomous logistic
regression}
\usage{
ehpoly(fit, z)
}
\arguments{
\item{fit}{a model fit from \code{multinom}}

\item{z}{design matrix for correspondence between disease subtypes and
tumor markers. Number of rows should equal number of subtypes, number of
columns should equal number of tumor markers. -1 indicates tumor marker
positive within a subtype, 1 indicates tumor marker negative within a
subtype.}
}
\value{
Returns a list. \code{beta} is the estimated parameters from
the fit. \code{beta_se} is estimates of the standard
errors of the parameters from the fit. \code{eh_pval}
is a vector of the etiologic heterogeneity p-values. \code{or_ci_p}
is a dataframe with a odds ratio (95% CI) for each risk factor/subtype
combination, as well as a column of etiologic heterogeneity p-values.
\code{beta_p} is a dataframe with the estimated parameters from the
 fit as well as a column of etiologic heterogeneity p-values.
\code{gamma_p} is a dataframe with estimates of the tumor marker effects and
their associated p-values.
}
\description{
#' \code{ehpoly} takes a multinomial logistic regression model fit and returns
results related to the question of whether a risk factor differs across
levels of a disease subtype, and whether a risk factor differs across levels
of each individual tumor marker or which the disease subtypes are comprised.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}

