% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstat_bin.R
\name{fstat_bin}
\alias{fstat_bin}
\title{Function to calculate the test statistic and proportion of variation
explained using MDMR}
\usage{
fstat_bin(Y, X)
}
\arguments{
\item{Y}{an NxK matrix of tumor marker data}

\item{X}{an NxP matrix of risk factor data}
}
\value{
returns a list containing f, the test statistic, and R, the
proportion of variation explained
}
\description{
\code{fstat} Computes the test statistic and proportion of
variation explained using multivariate distance matrix regression. For binary
tumor marker data. Uses asymmetric binary distance metric.
}
\references{
Zapala MA, Schork NJ. Multivariate regression analysis of
distance matrices for testing associations between gene expression patterns
and related variables. PNAS 2006; 103(51):19430-35.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
