% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmc.R
\name{minmc}
\alias{minmc}
\title{Calculate the minimum misclassification rate in a cross-tabulation}
\usage{
minmc(tab)
}
\arguments{
\item{tab}{a cross-tabulation table of a class solution from k-means
clustering to the true class solution}
}
\description{
\code{minmc} provides a way of reconciling the arbitrary class
labels that result from k-means clustering to obtain the minimum
misclassification rate of a k-means clustering result to a true class
solution, after aligning the class labels
}
\examples{

# Example where there is perfect alignment of class results, but the class
# labels do not correspond and need to be aligned
trueclass <- rep(c(1, 2, 3, 4), times = 20)
kclass <- rep(c(3, 2, 1, 4), times = 20)
table(kclass, trueclass)
minmc(table(kclass, trueclass))

}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
