% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plrsub.R
\name{plrsub}
\alias{plrsub}
\title{Conduct an analysis of etiologic heterogeneity for pre-defined subtypes
using polytomous logistic regression}
\usage{
plrsub(cls, m, rf, data, weight = NULL)
}
\arguments{
\item{cls}{the name of the variable in the data that contains numeric class
labels. This should be 0 for all controls. Argument must be supplied in
quotes.}

\item{m}{is the number of subtypes}

\item{rf}{a list of the names of the binary or continuous risk factors.
For binary risk factors the lowest level will be used as the reference level.}

\item{data}{the name of the dataframe that contains the tumor markers and risk
factors.}

\item{weight}{an optional argument to weights in the \code{mlogit} call.
Defaults to NULL.}
}
\value{
Returns a list.

\code{beta} is a matrix containing the estimated beta parameters with a
column for each risk factor and a row for each disease subtype.

\code{beta_se} contains the associated standard errors.

\code{eh_pval} is a vector of p-values for testing whether each risk factor
differs across levels of the disease subtype.

\code{or_ci_p} is a dataframe with a odds ratio (95% CI) for each risk
factor/subtype combination, as well as a column of etiologic heterogeneity
p-values.

\code{beta_se_p} is a dataframe with the estimated beta parameters (SE) for
each risk factor/subtype combination, as well as a column of
etiologic heterogeneity p-values.
}
\description{
\code{plrsub} takes a vector of class labels for pre-specified
subtypes and a list of risk factors and returns results related to the
question of whether each risk factor differs across levels of the disease
subtypes

Input is a dataframe that contains the risk factors of interest and a
variable containing numeric class labels that is 0 for control subjects.
Risk factors can be either binary or continuous. For categorical risk
factors, a reference level should be selected and then indicator variables
for each remaining level of the risk factor should be created.
Categorical risk factors entered as is will be treated as ordinal.
Class labels for the cases can be specified as a vector. The multinomial
logistic regression model is fit using the \code{mlogit} function from the
\code{mlogit} package.
}
\examples{

# generate the data - 4 classes and 2 risk factors
class <- rep(c(0, 1, 2, 3, 4), times = c(1000, 250, 250, 250, 250))
x <- matrix(rnorm(2000 * 2), 2000, 2) +
model.matrix(~factor(class))[, -1] \%*\%
t(matrix(c(1.5, 0, 0.75, 0.25, 0.25, 0.75, 0, 1.5), ncol = 4))
df <- data.frame(class = class, x1 = x[, 1], x2 = x[, 2])

plrsub("class", 4, c("x1", "x2"), df)

}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
