% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plrtm.R
\name{plrtm}
\alias{plrtm}
\title{Conduct an analysis of etiologic heterogeneity for individual tumor markers
using polytomous logistic regression}
\usage{
plrtm(tm, rf, case, data)
}
\arguments{
\item{tm}{a list of the names of the binary tumor markers.
Each must have levels 0 or 1 for case subjects. This value will be missing
for all control subjects.}

\item{rf}{a list of the names of the binary or continuous risk factors.
For binary risk factors the lowest level will be used as the reference level.}

\item{case}{denotes the variable that contains each subject's status as a
case or control. This value should be 1 for cases and 0 for controls.
Argument must be supplied in quotes.}

\item{data}{the name of the dataframe that contains the tumor markers and risk
factors.}
}
\value{
Returns a list.

\code{beta} is a matrix containing the estimated beta parameters with a
column for each risk factor and a row for each disease subtype.

\code{beta_se} contains the associated standard errors.

\code{eh_pval} is a vector of p-values for testing whether each risk factor
differs across levels of the disease subtype.

\code{gamma} is a matrix containing the estimated gamma parameters,
obtained as linear combinations of the beta parameters with a column for each
risk factor and a row for each tumor marker.

\code{gamma_se} contains the associated standard errors.

\code{gamma_p} is a matrix of p-values for testing whether each risk factor
differs across levels of each tumor marker, with a column for each risk
factor and a row for each tumor marker.

\code{or_ci_p} is a dataframe with a odds ratio (95% CI) for each risk
factor/subtype combination, as well as a column of etiologic heterogeneity
p-values.

\code{beta_se_p} is a dataframe with the estimated beta parameters (SE) for
each risk factor/subtype combination, as well as a column of
etiologic heterogeneity p-values.

\code{gamma_se_p} is a dataframe with estimates of the gamma tumor marker
effects (SE) and their associated p-values.
}
\description{
\code{plrtm} takes a list of individual tumor markers and a
list of risk factors and returns results related to the question of whether
each risk factor differs across levels of the disease subtypes and the
question of whether each risk factor differs across levels of each individual
tumor marker of which the disease subtypes are comprised.

Input is a dataframe that contains the individual tumor markers, the risk
factors of interest, and an indicator of case or control status.
The tumor markers must be binary and must have levels
0 or 1 for cases. The tumor markers should be left missing for control
subjects. For categorical tumor markers, a reference level should be selected
and then indicator variables for each remaining level of the tumor marker
should be created. For continuous tumor markers, categories should be formed
and then indicator variables can be constructed as in the case of categorical
tumor markers. Risk factors can be either binary or continuous. For
categorical risk factors, a reference level should be selected and then
indicator variables for each remaining level of the risk factor should be
created. Categorical risk factors entered as is will be treated as ordinal.
}
\examples{

# generate the data - 2 tumor markers and 2 risk factors
class <- rep(c(0, 1, 2, 3, 4), times = c(1000, 250, 250, 250, 250))
tm1 <- c(rep(0, 1000), rep(c(0, 1, 0, 1), each = 250))
tm2 <- c(rep(0, 1000), rep(c(0, 1), each = 500))
x <- matrix(rnorm(2000 * 2), 2000, 2) +
model.matrix(~factor(class))[, -1] \%*\%
t(matrix(c(1.5, 0, 0.75, 0.25, 0.25, 0.75, 0, 1.5), ncol = 4))
df <- data.frame(class = class, x1 = x[, 1], x2 = x[, 2], tm1 = tm1, tm2 = tm2)
df$caseind <- ifelse(df$class == 0, 0, 1)

plrtm(c("tm1", "tm2"), c("x1", "x2"), "caseind", df)


}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
