% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_factor_test.R
\name{posthoc_factor_test}
\alias{posthoc_factor_test}
\title{Post-hoc test to obtain overall p-value for a factor variable used in a
\code{eh_test_subtype} fit.}
\usage{
posthoc_factor_test(fit, factor, nlevels)
}
\arguments{
\item{fit}{the resulting \code{eh_test_subtype} fit.}

\item{factor}{is the name of the factor variable of interest, supplied
in quotes, e.g. \code{factor = "race"}. Only supports a single factor.}

\item{nlevels}{is the number of levels the factor variable in \code{factor}
has.}
}
\value{
Returns a list.

\code{pval} is a formatted p-value.

\code{pval_raw} is the raw, unformatted p-value.
}
\description{
\code{posthoc_factor_test} takes a \code{eh_test_subtype} fit
and returns an overall p-value for a specified factor variable.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
