% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueDsim.R
\name{trueDsim}
\alias{trueDsim}
\title{Estimate the true population D}
\usage{
trueDsim(N, M, pi, P, mu_m)
}
\arguments{
\item{N}{population sample size}

\item{M}{number of disease subtypes}

\item{pi}{vector of control and subtype prevalences
i.e. c(pi0, pi1, pi2, pi3) for controls and 3 subtypes}

\item{P}{number of risk factors}

\item{mu_m}{P x M matrix of risk factor means for the cases, by default all
risk factors have mean 0 for control subjects}
}
\description{
Adapted from code written by Venkat Seshan and made
generalizable to any number of subtypes, any number of risk factors, and
any risk factor means for the cases
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
