% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstar.R
\name{dstar}
\alias{dstar}
\title{Estimate the incremental explained risk variation in a case-only study}
\usage{
dstar(formula, label, M, data)
}
\arguments{
\item{formula}{an mFormula() model formula for a polytmous logistic
regression model to be fit with mlogit() using the appropriate variable
names from the data of interest.}

\item{label}{the name of the subtype variable in the data with values 1
through M. Should be supplied in quotes, e.g. label = "subtype".
The highest frequency subtype will be used as the reference subtype in
model}

\item{M}{the number of subtypes among case subjects. For M>=2.}

\item{data}{the name of the dataframe that contains the relevant variables}
}
\description{
\code{dstar} estimates the incremental explained risk variation
across a set of pre-specified disease subtypes in a case-only study.
This function takes a model formula and a wide dataset and does the needed
transformation on the dataset to get the correct format, fits the polytomous
logistic regression model, and calculates D based on the resulting risk
predictions.
}
\examples{

# specify the model formula for three risk factors, x1, x2 and x3
mform <- mlogit::mFormula(subtype ~ 1 | x1 + x2 + x3)

# calculate D*
# Exclude controls from data as this is a case-only calculation
dstar(mform, "subtype", 3, subtype_data[subtype_data$subtype > 0, ])

}
\references{
Begg, C. B., Seshan, V. E., Zabor, E. C., Furberg, H., Arora, A.,
Shen, R., . . . Hsieh, J. J. (2014). Genomic investigation of etiologic
heterogeneity: methodologic challenges. BMC Med Res Methodol, 14, 138.
doi: 10.1186/1471-2288-14-138
}
