% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehCalcD.R
\name{ehCalcD}
\alias{ehCalcD}
\title{Function to calculate the D metric in a case/control setting}
\usage{
ehCalcD(data, cls, k, formula)
}
\arguments{
\item{data}{a data frame with the covariates}

\item{cls}{the class variable with values 0 through k, where 0 is for
control subjects and 1:k are labels for the subtypes}

\item{k}{is the number of classes}

\item{formula}{is the model formula for the polytomous logistic regression
model, with "resp" on the lhs and the covariates of interest, located in
\code{data} on the rhs}
}
\value{
returns a 3-digit numeric value
}
\description{
\code{ehCalcD} takes a data set and information about class
membership and calculates the D metric.
}
\references{
Begg CB, Zabor EC, Bernstein JL, Bernstein L,
Press MF, Seshan VE. A conceptual and methodological framework
for investigating etiologic heterogeneity. Stat Med 2013; 32(29):5039-52.
}
\author{
Emily C Zabor \email{zabore@mskcc.org}
}
